Factorization: Add Timestampable trait to entities
This commit adds the Timestampable trait to manage timestamps
in the application's entities. Entities now have createdAt and
updatedAt properties along with corresponding methods for
automatically updating timestamps.
