+++ src/Entity/Theme.php
@@ -3,15 +3,13 @@
 namespace App\Entity;
 
 use App\Repository\ThemeRepository;
-use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
-use Gedmo\Mapping\Annotation as Gedmo;
-
-
+use App\Trait\TimestampableTrait;
 
 #[ORM\Entity(repositoryClass: ThemeRepository::class)]
 class Theme
 {
+    use TimestampableTrait;
     public const SIDEBAR_POSITION_LEFT = 'left';
     public const SIDEBAR_POSITION_RIGHT = 'right';
 
@@ -43,14 +41,6 @@ class Theme
     #[ORM\ManyToOne]
     #[ORM\JoinColumn(nullable: false)]
     private ?Font $subtitle_font = null;
-
-    #[Gedmo\Timestampable(on: 'create')]
-    #[ORM\Column(name:'created_at', type: Types::DATETIME_MUTABLE)]
-    private $created_at;
-
-    #[Gedmo\Timestampable(on: 'update')]
-    #[ORM\Column(name: 'updated_at', type: Types::DATETIME_MUTABLE)]
-    private $updated_at;
 
     #[ORM\Column(type: 'string', length: 10)]
     private ?string $sidebar_position = null;
@@ -146,30 +136,6 @@ class Theme
     public function setSubtitleFont(?Font $subtitle_font): static
     {
         $this->subtitle_font = $subtitle_font;
-
-        return $this;
-    }
-
-    public function getCreatedAt(): ?\DateTimeImmutable
-    {
-        return $this->created_at;
-    }
-
-    public function setCreatedAt(\DateTimeImmutable $created_at): static
-    {
-        $this->created_at = $created_at;
-
-        return $this;
-    }
-
-    public function getUpdatedAt(): ?\DateTimeInterface
-    {
-        return $this->updated_at;
-    }
-
-    public function setUpdatedAt(\DateTimeInterface $updated_at): static
-    {
-        $this->updated_at = $updated_at;
 
         return $this;
     }
